Cleanup of HazelcastStarter package
* fixed some smaller issues and warnings
* fixed usage of wrong rethrow() method in
  AbstractEventJournalBounceTest
* renamed rethrow() method in Utils to rethrowGuardianException() to
  ease static import of ExceptionUtil.rethrow()
* renamed Utils to HazelcastStarterUtils to make it less attractive
  in class import dialogs
* added missing tests for HazelcastStarter classes

+++ hazelcast/src/test/java/com/hazelcast/test/starter/DataAwareEntryEventConstructor.java
@@ -28,8 +28,7 @@ public class DataAwareEntryEventConstructor extends AbstractStarterObjectConstru
     }
 
     @Override
-    Object createNew0(Object delegate)
-            throws Exception {
+    Object createNew0(Object delegate) throws Exception {
         // locate required classes on target class loader
         ClassLoader starterClassLoader = targetClass.getClassLoader();
         Class<?> dataClass = starterClassLoader.loadClass("com.hazelcast.nio.serialization.Data");
@@ -49,15 +48,17 @@ public class DataAwareEntryEventConstructor extends AbstractStarterObjectConstru
         Object dataMergingValue = getFieldValueReflectively(delegate, "dataMergingValue");
 
         Object[] args = new Object[]{
-                member, eventTypeId.intValue(), source,
-                dataKey, dataNewValue,
-                dataOldValue, dataMergingValue,
+                member,
+                eventTypeId,
+                source,
+                dataKey,
+                dataNewValue,
+                dataOldValue,
+                dataMergingValue,
                 serializationService,
         };
 
         Object[] proxiedArgs = proxyArgumentsIfNeeded(args, starterClassLoader);
-
         return constructor.newInstance(proxiedArgs);
     }
-
 }
